add test event state
